[readme] Use gradlew script to bootstrap gradle on Unix
Since GitHub made it really difficult to comment on a rich preview, I am filing this more as an RFC.

`gradle assemble` doesn't seem to be the right command on either of my linux boxes, so I was wondering if this is a better approach. Or should system gradle be expected?